don't mutate the new resource
this line violates the rule that we never mutate the new resource.

in Chef 12 this causes real problems because if a file resource is
notified (run twice) then the rendered content is loaded into the
new resource and if the file resource content changes (i.e. its a
template with logic which changes when its notified) then it will
cause a failure because the rendered content will not match the
'requested' checksum.  where the 'requested' checksum is actually
the checksum loaded by this line.

fundamentally we have three different states that we're trying to
track:

- current state
- initial state
- requested state

the removed line tries to use the @new_resource for reporting initial
state, which then bleeds over into what chef thinks is requested state
in the next invokation.

without constructing a third @initial_resource and using that for
resource reporting comparison we can't solve this problem "right".

the tradeoff is we either break reporting here or break chef-client
runs.  this patch sacrifices reporting in order to make chef-client
work.